move non-disposable file handles implementations to the private namespace
